convert back to searching for original idea
Revert to search for Elvis. The 5-letter Pokemon Go character I tried occurred too often. (So much for trying to be relevant to the times.) The difference between that 5-letters and "ELVIS" must have something to do with either the frequency of synomous codons or the codons that happen to spell out "RALTS" matching better to a random distribution.